Fixed warnings about printf format in funcs primme_valloc and primme_calloc. Fixed usage of real workspace as integer in restart_locking, it caused problems in 32bits (really?). Increased the tolerance in checking the residual norm to pass some tests. Fixed issue in driver using a c++ compiler.

+++ PRIMMESRC/COMMONSRC/primme_interface.c
@@ -152,7 +152,7 @@ void *primme_valloc(size_t byteSize, const char *target) {
       perror("primme_alloc");
       fprintf(stderr,
          "ERROR(primme_alloc): Could not allocate %lu bytes for: %s\n",
-         byteSize, target);
+         (unsigned long int)byteSize, target);
       fflush(stderr);
       exit(EXIT_FAILURE);
    }
@@ -170,7 +170,7 @@ void *primme_calloc(size_t nelem, size_t elsize, const char *target) {
       perror("primme_calloc");
       fprintf(stderr, 
          "ERROR(primme_calloc): Could not allocate %lu elements of %lu bytes for: %s\n",
-         nelem, elsize, target);
+         (unsigned long int)nelem, (unsigned long int)elsize, target);
       fflush(stderr);
       exit(EXIT_FAILURE);
    }
